+++ usage/WebApp/Controllers/MainController.cs
@@ -5,7 +5,7 @@ using MindSphereSdk.Core.Common;
 using MindSphereSdk.Core.EventManagement;
 using MindSphereSdk.Core.IotTimeSeries;
 using MindSphereSdk.Core.IotTsAggregates;
-using Newtonsoft.Json;
+using MindSphereSdk.Core.Serialization;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -351,11 +351,11 @@ namespace WebApp.Controllers
             var iotClient = _sdk.GetIotTimeSeriesClient();
             var request = new GetTimeSeriesRequest()
             {
-                EntityId = "ec206f76b04a49a4938c1573b35b6688",
-                PropertySetName = "acceleration",
+                EntityId = "da4aabbd3f2f488da7ef75fa506a8eaa",
+                PropertySetName = "aspect1",
                 From = DateTime.Now.AddDays(-1),
                 To = DateTime.Now,
-                Limit = 10
+                Limit = 100
             };
             List<TestTimeSeriesData> timeSeries = (await iotClient.GetTimeSeriesAsync<TestTimeSeriesData>(request)).ToList();
             
@@ -370,9 +370,13 @@ namespace WebApp.Controllers
 
             List<TestTimeSeriesData> timeSeriesData = new()
             {
-                new TestTimeSeriesData(nowUtc, 0.5, 0.7, 0.3),
-                new TestTimeSeriesData(nowUtc.AddMinutes(1), 0.8, 1.2, 0.7),
-                new TestTimeSeriesData(nowUtc.AddMinutes(2), 1.6, 0.2, 0.5)
+                //new TestTimeSeriesData(nowUtc, 0.5, 0.7, 0.3),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(1), 0.8, 1.2, 0.7),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(2), 1.6, 0.3, 0.5),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(3), 0.1, 1.5, 2.1),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(4), 1.4, 0.2, 1.4),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(5), 1.3, 0.4, 0.9),
+                //new TestTimeSeriesData(nowUtc.AddMinutes(6), 0.9, 0.8, 1.8),
             };
             //timeSeriesData.Add(new { _time = DateTime.Now, x = 0.5, y = 0.7, z = 0.3 });
             //timeSeriesData.Add(new { _time = DateTime.Now.AddMinutes(1), x = 0.8, y = 1.2, z = 0.7 });
@@ -382,8 +386,8 @@ namespace WebApp.Controllers
             {
                 new TimeSeriesSet()
                 {
-                    EntityId = "ec206f76b04a49a4938c1573b35b6688",
-                    PropertySetName = "acceleration",
+                    EntityId = "da4aabbd3f2f488da7ef75fa506a8eaa",
+                    PropertySetName = "aspect1",
                     Data = timeSeriesData
                 }
             };
@@ -401,13 +405,17 @@ namespace WebApp.Controllers
         public async Task<ActionResult> PutTimeSeries()
         {
             var iotClient = _sdk.GetIotTimeSeriesClient();
-            DateTime nowUtc = DateTime.Now.ToUniversalTime();
+            DateTime now = DateTime.Now;
 
             List<TestTimeSeriesData> timeSeriesData = new()
             {
-                new TestTimeSeriesData(nowUtc, 0.5, 0.7, 0.3),
-                new TestTimeSeriesData(nowUtc.AddMinutes(1), 0.8, 1.2, 0.7),
-                new TestTimeSeriesData(nowUtc.AddMinutes(2), 1.6, 0.2, 0.5)
+                new TestTimeSeriesData(now, 0.5),
+                new TestTimeSeriesData(now.AddMinutes(1), 0.8),
+                //new TestTimeSeriesData(now.AddMinutes(2), 1.6, 0.3, 0.5),
+                //new TestTimeSeriesData(now.AddMinutes(3), 0.1, 1.5, 2.1),
+                //new TestTimeSeriesData(now.AddMinutes(4), 1.4, 0.2, 1.4),
+                //new TestTimeSeriesData(now.AddMinutes(5), 1.3, 0.4, 0.9),
+                //new TestTimeSeriesData(now.AddMinutes(6), 0.9, 0.8, 1.8),
             };
             //timeSeriesData.Add(new { _time = DateTime.Now, x = 0.5, y = 0.7, z = 0.3 });
             //timeSeriesData.Add(new { _time = DateTime.Now.AddMinutes(1), x = 0.8, y = 1.2, z = 0.7 });
@@ -416,8 +424,8 @@ namespace WebApp.Controllers
             PutTimeSeriesRequest request = new()
             {
                 Data = timeSeriesData,
-                EntityId = "ec206f76b04a49a4938c1573b35b6688",
-                PropertySetName = "acceleration"
+                EntityId = "da4aabbd3f2f488da7ef75fa506a8eaa",
+                PropertySetName = "aspect1"
             };
             await iotClient.PutTimeSeriesAsync(request);
 
@@ -511,10 +519,10 @@ namespace WebApp.Controllers
         {
             var assetClient = _sdk.GetAssetManagementClient();
 
-            var newAspectType = new AspectTypeUpdate()
+            var newAspectType = new AspectTypeAddUpdate()
             {
-                Name = "My_new_asset",
-                Category = "static",
+                Name = "Dotnet_sdk",
+                Category = "dynamic",
                 Description = "Test",
                 Scope = "private",
                 Variables = new VariableDetail[] {
@@ -529,7 +537,7 @@ namespace WebApp.Controllers
 
             var request = new PutAspectTypeRequest()
             {
-                Id = "iiotdgli.My_new_asset",
+                Id = "prsdevex.Dotnet_sdk",
                 IfNoneMatch = "*",
                 AspectType = newAspectType
             };
@@ -543,7 +551,7 @@ namespace WebApp.Controllers
         {
             var assetClient = _sdk.GetAssetManagementClient();
 
-            var updatedAspectType = new AspectTypeUpdate()
+            var updatedAspectType = new AspectTypeAddUpdate()
             {
                 Name = "My_new_aspect_type",
                 Category = "static",
@@ -575,7 +583,7 @@ namespace WebApp.Controllers
         {
             var assetClient = _sdk.GetAssetManagementClient();
 
-            var updatedAspectType = new AspectTypeUpdate()
+            var updatedAspectType = new AspectTypeAddUpdate()
             {
                 Description = "Patched test"
             };
@@ -614,7 +622,7 @@ namespace WebApp.Controllers
             var assetClient = _sdk.GetAssetManagementClient();
             var request = new GetAssetTypeRequest()
             {
-                Id = "core.basicagent"
+                Id = "prsdevex.Dotnet_sdk"
             };
             AssetType assetType = await assetClient.GetAssetTypeAsync(request);
 
@@ -659,7 +667,7 @@ namespace WebApp.Controllers
 
             var aspects = new List<AspectPut>
             {
-                new AspectPut() { Name = "acceleration", AspectTypeId = "iiotdgli.acceleration" }
+                new AspectPut() { Name = "aspect1", AspectTypeId = "prsdevex.Dotnet_sdk" }
             };
 
             var newAssetType = new AssetTypeUpdate()
@@ -674,7 +682,7 @@ namespace WebApp.Controllers
 
             var request = new PutAssetTypeRequest()
             {
-                Id = "iiotdgli.My_new_asset_type",
+                Id = "prsdevex.Dotnet_sdk",
                 IfMatch = "0",
                 AssetType = newAssetType
             };
@@ -690,13 +698,16 @@ namespace WebApp.Controllers
 
             var updatedAssetType = new AssetTypeUpdate()
             {
-                Description = "Patched asset type"
-            };
+                Aspects = new List<AspectPut>
+                {
+                    new AspectPut() { Name = "aspect1", AspectTypeId = "prsdevex.Dotnet_sdk" }
+                }
+         };
 
             var request = new PatchAssetTypeRequest()
             {
-                Id = "iiotdgli.My_new_asset_type",
-                IfMatch = "1",
+                Id = "prsdevex.Dotnet_sdk",
+                IfMatch = "2",
                 AssetType = updatedAssetType
             };
             var assetType = await assetClient.PatchAssetTypeAsync(request);
@@ -937,63 +948,55 @@ namespace WebApp.Controllers
 
     public class TestTimeSeriesData
     {
-        [JsonProperty("_time")]
+        [MindSphereName("_time")]
         public DateTime Time { get; set; }
 
-        [JsonProperty("x")]
+        [MindSphereName("velocity")]
         public double X { get; set; }
 
-        [JsonProperty("y")]
-        public double Y { get; set; }
-
-        [JsonProperty("z")]
-        public double Z { get; set; }
-
-        public TestTimeSeriesData(DateTime time, double x, double y, double z)
+        public TestTimeSeriesData(DateTime time, double v)
         {
             Time = time;
-            X = x;
-            Y = y;
-            Z = z;
+            X = v;
         }
     }
 
     public class TestAggregateTsData : AggregateSet
     {
-        [JsonProperty("x")]
+        [MindSphereName("x")]
         public AggregateVariable X { get; set; }
 
-        [JsonProperty("y")]
+        [MindSphereName("y")]
         public AggregateVariable Y { get; set; }
 
-        [JsonProperty("z")]
+        [MindSphereName("z")]
         public AggregateVariable Z { get; set; }
     }
 
     public class VariableMap
     {
-        [JsonProperty("var4")]
+        [MindSphereName("var4")]
         public VariableUpdate Var3 { get; set; }
 
-        [JsonProperty("var3")]
+        [MindSphereName("var3")]
         public VariableUpdate Var4 { get; set; }
     }
 
     public class MyEventAddUpdate : EventAddUpdate
     {
-        [JsonProperty("description")]
+        [MindSphereName("description")]
         public string Description { get; set; }
 
-        [JsonProperty("severity")]
+        [MindSphereName("severity")]
         public int Severity { get; set; }   
     }
 
     public class MyEvent : Event
     {
-        [JsonProperty("description")]
+        [MindSphereName("description")]
         public string Description { get; set; }
 
-        [JsonProperty("severity")]
+        [MindSphereName("severity")]
         public int Severity { get; set; }
     }
 }